implemented user data access check, users will no longer be able to access data they are not owner of ( when trying to make requests directly to the backend with a valid Bearer)

+++ server/controllers/personality-controller.js
@@ -40,7 +40,7 @@ const httpCreatePersonality = async (req, res) => {
 };
 
 const httpGetPersonalities = async (req, res) => {
-  const result = await chatPersonalityModel.getPersonalityDetails();
+  const result = await chatPersonalityModel.getPersonalityDetails(req.body.requestedbyUser);
 
   if (result === false) {
     res.status(500).json({});
@@ -50,6 +50,11 @@ const httpGetPersonalities = async (req, res) => {
 };
 
 const httpUpdatePersonality = async (req, res) => {
+  if(await chatPersonalityModel.isAuthorized(req.params.id, req.body.requestedbyUser) === false){
+    res.status(401).json({ error: "User is not authorized to access data"});
+    return false;
+  }
+
   if (!req.body.name || typeof req.body.name !== "string") {
     res.status(400).json({ error: "Must provide a name string" });
     return false;
@@ -89,6 +94,11 @@ const httpUpdatePersonality = async (req, res) => {
 };
 
 const httpDeletePersonality = async (req, res) => {
+  if(await chatSessionModel.isAuthorized(req.params.id, req.body.requestedbyUser) === false){
+    res.status(401).json({ error: "User is not authorized to access data"});
+    return false;
+  }
+
   const result = await chatPersonalityModel.deletePersonality(req.params.id);
   if (result === false) {
     res.status(500).json({});
